Add some sysdeps needed for bash to get a prompt

+++ sysdeps/obos/include/abi-bits/pid_t.h
@@ -2,7 +2,7 @@
 #ifndef _ABIBITS_PID_T_H
 #define _ABIBITS_PID_T_H
 
-typedef long unsigned pid_t;
+typedef int pid_t;
 
 #endif /* _ABIBITS_PID_T_H */
